requirements.txt
openai
azure.cognitiveservices.speech
pygame
pydub
mutagen
pillow